+++ src/components/number.js
@@ -1,4 +1,7 @@
 var fs = require('fs');
+var createNumberMask = require('text-mask-addons').createNumberMask;
+var _get = require('lodash/get');
+var _isNil = require('lodash/isNil');
 
 module.exports = function(app) {
   app.config([
@@ -10,6 +13,29 @@ module.exports = function(app) {
       formioComponentsProvider.register('number', {
         title: 'Number',
         template: 'formio/components/number.html',
+        tableView: function(data, options) {
+          if (_isNil(data)) {
+            return '';
+          }
+
+          var separators = options.util.getNumberSeparators();
+          var component = options.component;
+          var decimalLimit = options.util.getNumberDecimalLimit(component);
+
+          if (!component.delimiter) {
+            separators.delimiter = '';
+          }
+
+          return options.util.formatNumber(data, createNumberMask({
+            prefix: '',
+            suffix: '',
+            thousandsSeparatorSymbol: _get(component, 'thousandsSeparator', separators.delimiter),
+            decimalSymbol: _get(component, 'decimalSymbol', separators.decimalSeparator),
+            decimalLimit: _get(component, 'decimalLimit', decimalLimit),
+            allowNegative: _get(component, 'allowNegative', true),
+            allowDecimal: _get(component, 'allowDecimal', !(component.validate && component.validate.integer))
+          }));
+        },
         settings: {
           autofocus: false,
           input: true,