+++ src/components/radio.js
@@ -7,10 +7,10 @@ module.exports = function(app) {
       formioComponentsProvider.register('radio', {
         title: 'Radio',
         template: 'formio/components/radio.html',
-        tableView: function(data, component) {
-          for (var i in component.values) {
-            if (component.values[i].value === data) {
-              return component.values[i].label;
+        tableView: function(data, options) {
+          for (var i in options.component.values) {
+            if (options.component.values[i].value === data) {
+              return options.component.values[i].label;
             }
           }
           return data;