+++ src/components/survey.js
@@ -8,14 +8,14 @@ module.exports = function(app) {
         title: 'Survey',
         template: 'formio/components/survey.html',
         group: 'advanced',
-        tableView: function(data, component) {
+        tableView: function(data, options) {
           var view = '<table class="table table-striped table-bordered"><thead>';
           var values = {};
-          angular.forEach(component.values, function(v) {
+          angular.forEach(options.component.values, function(v) {
             values[v.value] = v.label;
           });
           if (data) {
-            angular.forEach(component.questions, function(question) {
+            angular.forEach(options.component.questions, function(question) {
               view += '<tr>';
               view += '<th>' + question.label + '</th>';
               view += '<td>' + values[data[question.value]] + '</td>';