+++ src/components/table.js
@@ -25,26 +25,31 @@ module.exports = function(app) {
           condensed: false,
           tableView: false
         },
-        tableView: function(data, component, $interpolate, componentInfo, tableChild) {
+        tableView: function(data, options) {
           var view = '<table class="table table-striped table-bordered table-child">';
 
-          if (!tableChild) {
+          if (!options.tableChild) {
             view += '<thead>';
-            view += '<tr><th>Table (' + component.key + ')</th></tr>';
+            view += '<tr><th>Table (' + options.component.key + ')</th></tr>';
             view += '</thead>';
           }
 
           view += '<tbody>';
 
-          for (var row = 0; row < component.numRows; row++) {
+          for (var row = 0; row < options.component.numRows; row++) {
             view += '<tr>';
-            for (var col = 0; col < component.numCols; col++) {
+            for (var col = 0; col < options.component.numCols; col++) {
               view += '<td>';
               // Each column is its own table.
               view += '<table class="table table-striped table-bordered table-child">';
               view += '<tbody>';
-              angular.forEach(component.rows[row][col].components, function(component) {
-                view += '<tr>' + GridUtils.columnForComponent(data, component, $interpolate, componentInfo) + '</tr>';
+              angular.forEach(options.component.rows[row][col].components, function(component) {
+                view += '<tr>' + GridUtils.columnForComponent(data, {
+                  component: component,
+                  $interpolate: options.$interpolate,
+                  componentInfo: options.componentInfo,
+                  util: options.util
+                }) + '</tr>';
               });
 
               view += '</tbody></table>';